moving to data package database and adapter to ui
